requirements.txt
selenium==4
webdriver_manager
pytest
allure-pytest